pyproject.toml
[tool.poetry]
name = "case-conference"
version = "0.1.0"
description = ""
authors = ["sirendi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
openai = "^1.6.1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
fpdf = "^1.7.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"

[tool.poetry.group.lint.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
pydocstyle = "^6.3.0"
pylint = "^2.15.0"
isort = "^5.12.0"
types-requests = "^2.31.0.1"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
